Add debug line that kicks out the issue.
OK- here's what it is, it's the primes*2 iteration over known primes, at
memo.go:106. I noticed that 25 is in the "is prime" list, though 20
isn't. Running that loop for starting at 10, with prime = 5, you get:
  oddMax = 9
  oddMax % prime = 4
  oddMax - (oddMax % prime) = 5
And iterations up through 15, 25, 35, etc., which skips the 10s, which
is fine.

But what if p.max == 11?
  oddMax = 11
  oddMax % prime = 1
  oddMax - (oddMax % prime) = 10
And iterations up through 10, 20, 30, etc. which will mark odds one less
than the respective evens (9, 19, 29, etc), possibly incorrectly; and
which won't mark the '5s, likewise incorrectly.