Fix some locking discipline.
The thread launched by PrimesUpTo held the read lock for its entire life. If a
caller subsequently tried to use IsPrime on a larger number, the
invocation of computeUpTo would be blocked on the read lock... which
causes a deadlock.

So: copy out the listing while holding the read lock.